+++ NEWS.md
@@ -38,6 +38,7 @@ Version 0.17b0 - dev
 - BUGFIX: entity query did not match default values if the attribute was not present
 - BUGFIX: groupby query did not match default values if the attribute was not present
 - BUGFIX: ODAFC add-on - reintroduce accidentally removed global variable `exec_path` as `win_exec_path` 
+- BUGFIX: graphic entities are not allowed as `DICTIONARY` entries 
 
 Version 0.16.6 - 2021-08-28
 ---------------------------